main layout completed, layout
