+++ web/component/Header.js
@@ -3,8 +3,8 @@ export default function Header(props) {
   return (
     <div className={style.header_container}>
       <div className={style.header}>
-        <h1 className={style.headOne}>Matt</h1>
-        <h1 className={style.headTwo}>Atterholt</h1>
+        <h1 className={style.headOne}>Matt Atterholt</h1>
+        <h1 className={style.headTwo}>{props.pageName}</h1>
 
         <span className={style.bg_pattern}></span>
         <span className={style.bg_pattern}></span>
