+++ web/data/projectList.js
@@ -1,0 +1,40 @@
+const repos = [
+  {
+    title: `Work Hour Log`,
+    content:
+      "We worked in logging the hours with excel. Creating a front end web app to calculate the hours for the week. Primary language was JavaScript with REACT.js",
+    link: `https://github.com/matterholt/workHourLog`
+  },
+  {
+    title: `Python work scripts`,
+    content:
+      "Where I gotten my start and found the ways to create scripts to help with the task with our daily work !! review",
+    link: ``
+  },
+  {
+    title: `Abaqus Static curve script`,
+    content:
+      "A script that works with the Abaqus results odb. It has a library that is specific and only works with Python2",
+    link: `https://github.com/matterholt/abaqusStaticCurveCompiler`
+  },
+  {
+    title: `Weld Thickness Calculator`,
+    content:
+      "Simple calculator that calculates the thickness of the weld for a simulation",
+    link: `https://github.com/matterholt/WeldThicknessCalculator`
+  },
+  {
+    title: `Build Me`,
+    content:
+      "A project that is started up with a co-worker. We started this project to learn the full stack development. Where my primary role is to work on the backend",
+    link: `https://github.com/TEAM-BUILDME/tool-share-backend`
+  },
+  {
+    title: `FEA tracking system`,
+    content: `In the group I'm apart of information is passed using power point. A bug tracking program 
+      inspired me to build a web app. I hope to be able incorporate the function that a bug tracking application 
+      and apply it our work flow.`,
+    link: `https://github.com/matterholt/fea_tracker`
+  }
+];
+export default repos;
